cmd/scoring-service/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4